Finalized plotter
Warning: Due to this branch of the project not being within the scope of the area of study of the project, the training performance plotter file has been created using the help of AI models (ChatGPT, Gemini).

This week, I have made major improvements and changes to the Performance During Inference tab.

- The translation of the sample sentences can now be successfully generated. The file translate.py has been written based on hkproj's pytorch-transformer, translate.py file with modifications made.

- Have added two time-tracking variables:
   1. Decoding Time: Time taken for model to generate attention maps given source and target texts.
   2. Generation of Tokens Per Second: tracks the speed of the model during inference, translating the sample sentence.

The color scheme of the attention head heat maps have been changed for aesthetic effects